Chunter: FileBrowser - extract filters and sorting
