+++ plugins/attachment-assets/lang/en.json
@@ -26,6 +26,8 @@
     "FileBrowserSortOldest": "Oldest file",
     "FileBrowserSortAZ": "A to Z",
     "FileBrowserSortZA": "Z to A",
+    "FileBrowserSortSmallest": "Smallest file",
+    "FileBrowserSortBiggest": "Biggest file",
     "FileBrowserDateFilterAny": "Any time",
     "FileBrowserDateFilterToday": "Today",
     "FileBrowserDateFilterYesterday": "Yesterday",
