+++ plugins/attachment-resources/src/plugin.ts
@@ -37,18 +37,8 @@ export default mergeIds(attachmentId, attachment, {
     FileBrowserSortOldest: '' as IntlString,
     FileBrowserSortAZ: '' as IntlString,
     FileBrowserSortZA: '' as IntlString,
-    FileBrowserDateFilterAny: '' as IntlString,
-    FileBrowserDateFilterToday: '' as IntlString,
-    FileBrowserDateFilterYesterday: '' as IntlString,
-    FileBrowserDateFilter7Days: '' as IntlString,
-    FileBrowserDateFilter30Days: '' as IntlString,
-    FileBrowserDateFilter3Months: '' as IntlString,
-    FileBrowserDateFilter12Months: '' as IntlString,
-    FileBrowserTypeFilterAny: '' as IntlString,
-    FileBrowserTypeFilterImages: '' as IntlString,
-    FileBrowserTypeFilterAudio: '' as IntlString,
-    FileBrowserTypeFilterVideos: '' as IntlString,
-    FileBrowserTypeFilterPDFs: '' as IntlString,
+    FileBrowserSortSmallest: '' as IntlString,
+    FileBrowserSortBiggest: '' as IntlString,
     AddAttachmentToSaved: '' as IntlString,
     RemoveAttachmentFromSaved: '' as IntlString
   },
